# Performance Improvements to the Inventory Dashboard (`/material-inventory`)