--- a/application/views/viewMaterialInventory.ejs
+++ b/application/views/viewMaterialInventory.ejs
@@ -116,12 +116,12 @@
           </div>
           <div class='col col-left'>
             <span>Actual</span>
-            <h2><%= materialInventory.lengthOfMaterialInStock %></h2>
+            <h2 class='material-length-in-stock'><%= materialInventory.lengthOfMaterialInStock %></h2>
           </div>
           <div class='divide-line'></div>
           <div class='col col-right'>
             <span>Ordered</span>
-            <h2><%= materialInventory.lengthOfMaterialOrdered %></h2>
+            <h2 class='material-length-ordered'><%= materialInventory.lengthOfMaterialOrdered %></h2>
           </div>
         </div>