# Description

It was noticed that the inventory dashboard took a couple seconds to load with the minmal data that it was displaying.

After investigation, it was found that backend query methods were taking 1,760 ms (1.760 seconds) to execute multiple synchronous HTTP requests.

However, instead of performing multiple, I was able to reduce the number of HTTP requests sent down to around 1 and then process that data in code on the server side much quicker.

After the updates in this PR, the `1760ms` load time is now `~90ms` and the page functions identically the same as it did before 🎊 

This PR closes #156 

## Verification
![image](https://user-images.githubusercontent.com/42784674/201445399-6c88c866-7995-4590-b80d-87360032d81b.png)
> Before creating a Purchase Order of 12 rolls of 12 feet each (total of 144 feet)

![image](https://user-images.githubusercontent.com/42784674/201445476-66553f4e-76ec-423e-a889-b36c32b6e2d6.png)
> After creating a cloth purchase order of 12 rolls, each 12 feet (total of 144 feet). This purchase order was not marked as delivered yet

![image](https://user-images.githubusercontent.com/42784674/201445523-5ae57236-f769-4e10-84cf-0d671464e111.png)

> After marking the previously mentioned purchase order as delivered